WIP refactor send document notifications
