+++ packages/subscriptions/lib/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   Subscriptions: require('./Subscriptions'),
+  Document: require('./Document'),
   sendNotification: require('./send')
 }
