+++ servers/publikator/lib/Notifications.js
@@ -1,7 +1,6 @@
 const {
-  Subscriptions: {
-    getSubscriptionsForUserAndObject,
-    getSubscriptionsForUserAndObjects
+  Document: {
+    getSubscriptionsForDoc,
   }
 } = require('@orbiting/backend-modules-subscriptions')
 const { sendNotification } = require('@orbiting/backend-modules-subscriptions')
@@ -13,44 +12,6 @@ const {
   FRONTEND_BASE_URL
 } = process.env
 
-const getFormatSubscriptions = async (doc, formatDoc, context) => {
-  if (!formatDoc) {
-    return
-  }
-
-  return getSubscriptionsForUserAndObject(
-    null,
-    {
-      type: 'Document',
-      id: formatDoc.meta.repoId
-    },
-    context,
-    {
-      onlyEligibles: true
-    }
-  )
-}
-
-const getAuthorSubscriptions = (doc, context) => {
-  const { authorUserIds } = doc.meta
-  if (!authorUserIds) {
-    return
-  }
-
-  return getSubscriptionsForUserAndObjects(
-    null,
-    {
-      type: 'User',
-      ids: authorUserIds,
-      filter: 'Document'
-    },
-    context,
-    {
-      onlyEligibles: true
-    }
-  )
-}
-
 const getUsersWithSubscriptions = (subscriptions = [], { loaders }) => {
   return Promise.map(
     subscriptions,
@@ -61,29 +22,30 @@ const getUsersWithSubscriptions = (subscriptions = [], { loaders }) => {
   )
 }
 
-const notifyPublish = async (repoId, context, testUsers) => {
+const groupSubscribersByObjectId = (subscribers, key) => subscribers.reduce(
+  (agg, user) => {
+    const objectId = user.__subscription[key]
+    if (agg[objectId]) {
+      agg[objectId].push(user)
+    } else {
+      agg[objectId] = [user]
+    }
+    return agg
+  }, {}
+)
+
+const notifyPublish = async (repoId, context, testUser) => {
   const {
     loaders,
     t
   } = context
 
-  if (testUsers && !Array.isArray(testUsers)) {
+  if (testUser && !testUser.id) {
     throw new Error(t('api/unexpected'))
   }
 
   const doc = await loaders.Document.byRepoId.load(repoId)
   const docRepoId = doc.meta.repoId
-
-  const { repoId: formatRepoId } = getRepoId(doc.meta.format)
-  const formatDoc = formatRepoId && await loaders.Document.byRepoId.load(formatRepoId)
-
-  const [formatSubscriptions, authorsSubscriptions] = testUsers ? [] : await Promise.all([
-    getFormatSubscriptions(doc, formatDoc, context),
-    getAuthorSubscriptions(doc, context)
-  ])
-
-  const formatSubscribers = testUsers ||
-    await getUsersWithSubscriptions(formatSubscriptions, context)
 
   const eventInfo = {
     objectType: 'Document',
@@ -97,65 +59,90 @@ const notifyPublish = async (repoId, context, testUsers) => {
     type: 'discussion',
     tag: docRepoId
   }
-  let event
-  if (formatSubscribers.length) {
-    event = await sendNotification(
-      {
-        event: eventInfo,
-        users: formatSubscribers,
-        content: {
-          app: {
-            ...appContent,
-            title: t('api/notifications/doc/title', { title: inQuotes(formatDoc.meta.title) })
-          }
-        }
-      },
-      context
-    )
-  }
 
-  const notifiedUserIds = formatSubscribers.map(user => user.id)
-  const dedupedAuthorsSubscriptions = authorsSubscriptions
-    .filter(sub => !notifiedUserIds.includes(sub.userId))
-
-  if (dedupedAuthorsSubscriptions.length) {
-    const authorsSubscribers =
-      await getUsersWithSubscriptions(dedupedAuthorsSubscriptions, context)
-
-    const authorSubscribersByAuthorIds = authorsSubscribers.reduce(
-      (agg, user) => {
-        const { objectUserId: authorId } = user.__subscription
-        if (agg[authorId]) {
-          agg[authorId].push(user)
-        } else {
-          agg[authorId] = [user]
-        }
-        return agg
-      }, {}
-    )
-
-    await Promise.each(
-      Object.keys(authorSubscribersByAuthorIds),
-      async (authorId) => {
-        const author = await loaders.User.byId.load(authorId)
-        const subscribers = authorSubscribersByAuthorIds[authorId]
-
-        event = await sendNotification(
-          {
-            event: event ? { id: event.id } : eventInfo,
-            users: subscribers,
-            content: {
-              app: {
-                ...appContent,
-                title: t('api/notifications/doc/author/title', { name: author.name })
-              }
-            }
-          },
-          context
-        )
+  const docsSubscriptions = []
+  const authorsSubscriptions = []
+  await getSubscriptionsForDoc(
+    doc,
+    testUser?.id, //otherwise null => for all users
+    {
+      ...testUser ? {
+        simulate: true,
+        onlyEligibles: false,
+        includeNotActive: true,
+        uniqueUsers: false
+      } : {
+        onlyEligibles: true,
+        includeParents: true,
+        uniqueUsers: true,
       }
-    )
-  }
+    },
+    context
+  ).then( subs => subs.forEach( sub => {
+    if (sub.objectType === 'Document') {
+      docsSubscriptions.push(sub)
+    } else {
+      authorsSubscriptions.push(sub)
+    }
+  }))
+
+  let event
+
+  const docsSubscribers = await getUsersWithSubscriptions(docsSubscriptions, context)
+  const docSubscribersByDocId = groupSubscribersByObjectId(docsSubscribers, 'objectDocumentId')
+
+  await Promise.each(
+    Object.keys(docSubscribersByDocId),
+    async (docId) => {
+      const subscribedDoc = await loaders.Document.byRepoId.load(docId)
+      const subscribers = docSubscribersByDocId[docId]
+      console.log({type: 'doc', subscribers,
+        title: t('api/notifications/doc/title', { title: inQuotes(subscribedDoc.meta.title) })
+      })
+
+      event = await sendNotification(
+        {
+          event: event ? { id: event.id } : eventInfo,
+          users: subscribers,
+          content: {
+            app: {
+              ...appContent,
+              title: t('api/notifications/doc/title', { title: inQuotes(subscribedDoc.meta.title) })
+            }
+          }
+        },
+        context
+      )
+    }
+  )
+
+  const authorsSubscribers = await getUsersWithSubscriptions(authorsSubscriptions, context)
+  const authorSubscribersByAuthorId = groupSubscribersByObjectId(authorsSubscribers, 'objectUserId')
+
+  await Promise.each(
+    Object.keys(authorSubscribersByAuthorId),
+    async (authorId) => {
+      const author = await loaders.User.byId.load(authorId)
+      const subscribers = authorSubscribersByAuthorId[authorId]
+      console.log({type: 'author', subscribers,
+        title: t('api/notifications/doc/author/title', { name: author.name })
+      })
+
+      event = await sendNotification(
+        {
+          event: event ? { id: event.id } : eventInfo,
+          users: subscribers,
+          content: {
+            app: {
+              ...appContent,
+              title: t('api/notifications/doc/author/title', { name: author.name })
+            }
+          }
+        },
+        context
+      )
+    }
+  )
 }
 
 module.exports = {
